% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_offset.R
\name{distance_offset}
\alias{distance_offset}
\title{distance_offset}
\usage{
distance_offset(sf_point_data)
}
\arguments{
\item{sf_point_data}{point data}
}
\value{
returns column for distance point was offset by. If any below 50m, throws a warning and additional T/F "flag_within_50m" column
}
\description{
Add distance offset from original point location
}
\examples{
\dontrun{
  offset_points_within_boundary(
    sf_point_data = bcmaps::bc_cities(),
    sf_boundary = bcmaps::health_chsa(),
    sf_boundary_id_col = "cmnty_hlth_serv_area_code",
    sf_boundary_id_col_new_name = "chsa",
    sf_boundary_total_pop_col = "chsa_population_census"
    )
  }

}
\keyword{internal}
