% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_is_point_object.R
\name{is_point_sf}
\alias{is_point_sf}
\title{Stop Flags - Is Point Object}
\usage{
is_point_sf(sf_point_object)
}
\arguments{
\item{sf_point_object}{point object}
}
\value{
stops function if not a point/multipoint object
}
\description{
built in checks that stop function if not met
}
\examples{
\dontrun{
  offset_points_within_boundary(postal_code, bcmaps::health_chsa, "cmnty_hlth_serv_area_code", "chsa", "total_pop_chsa")
  }

}
\keyword{internal}
