% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_is_polygon_object.R
\name{is_polygon_sf}
\alias{is_polygon_sf}
\title{Stop Flags - Is Poylgon Object}
\usage{
is_polygon_sf(sf_polygon_object)
}
\arguments{
\item{sf_polygon_object}{polygon object}
}
\value{
stops function if not a polygon/multipolygon object
}
\description{
built in checks that stop function if not met
}
\examples{
\dontrun{
  offset_points_within_boundary(postal_code, bcmaps::health_chsa, "cmnty_hlth_serv_area_code", "chsa", "total_pop_chsa")
  }

}
\keyword{internal}
