% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offset_corrections.R
\name{offset_corrections}
\alias{offset_corrections}
\title{offset_corrections}
\usage{
offset_corrections(
  offset_data,
  offset_boundary_id_col_name_original,
  offset_boundary_id_col_name_offset,
  sf_boundary,
  sf_boundary_id_col,
  buffer_dist_to_correct_by_meters = 5
)
}
\arguments{
\item{offset_data}{sf object containing offset geometry needing review}

\item{offset_boundary_id_col_name_original}{boundary id col from original x/y coordinates in offset_data}

\item{offset_boundary_id_col_name_offset}{boundary id col from offset x/y coordinates in offset_data}

\item{sf_boundary}{sf boundary of interest}

\item{sf_boundary_id_col}{sf boundary id col desired to identify boundary assigned to point based on current geometry}

\item{buffer_dist_to_correct_by_meters}{buffer distance to be made within boundary}
}
\value{
returns an sf point object with adjusted offset points that now fall within their original polygon boundary
}
\description{
Correct Offset Point Positions

Used in offset_points_within_boundary(); corrects offset points that fell beyond their original boundary polygon
Have to review offset point geometry location and if any fall outside their original polygon location, nudge their placement back into original polygon
}
\examples{
\dontrun{
  offset_points_within_boundary(
    sf_point_data = bcmaps::bc_cities(),
    sf_boundary = bcmaps::health_chsa(),
    sf_boundary_id_col = "cmnty_hlth_serv_area_code",
    sf_boundary_id_col_new_name = "chsa",
    sf_boundary_total_pop_col = "chsa_population_census"
    )
  }

}
\keyword{internal}
