% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offset_point.R
\name{offset_point}
\alias{offset_point}
\title{offset_point}
\usage{
offset_point(
  sf_data_to_offset,
  crs_code = 3005,
  rand_dist_min = 50,
  rand_dist_max = 200,
  rand_angle_min = 1,
  rand_angle_max = 360
)
}
\arguments{
\item{sf_data_to_offset}{sf point object requiring offsetting}

\item{rand_dist_min}{default min = 50}

\item{rand_dist_max}{default max = 200}

\item{rand_angle_min}{default min = 1}

\item{rand_angle_max}{default max = 360}
}
\value{
returns an sf point object with offset points
}
\description{
Offset Coordinates

Used in offset_points_within_boundary()
Offset points to maintain confidentiality using the following offsetting methodology:
    x offset formula: x' = x0 + RandDist * cos(RandAngle)
    y offset formula: y' = y0 + RandDist * sin(RandAngle)
    Note: default min/max range for random distance number is between 50:200 and for random angle its between 1:360
Note: offsetting MUST be done in a crs fit for BC mapping like 3005 (pcs NAD 1983 BC Albers)

Note: offsetting is best done based off of a population density min/max handled in ave_dist_btw_ppl() and offset_points_within_boundary()
}
\examples{
\dontrun{
  offset_points_within_boundary(
    sf_point_data = bcmaps::bc_cities(),
    sf_boundary = bcmaps::health_chsa(),
    sf_boundary_id_col = "cmnty_hlth_serv_area_code",
    sf_boundary_id_col_new_name = "chsa",
    sf_boundary_total_pop_col = "chsa_population_census"
    )
  }

}
\keyword{internal}
