% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offset_points_within_boundary.R
\name{offset_points_within_boundary}
\alias{offset_points_within_boundary}
\title{offset_points_within_boundary}
\usage{
offset_points_within_boundary(
  sf_point_data,
  sf_boundary,
  sf_boundary_id_col,
  sf_boundary_id_col_new_name = "boundaryid",
  sf_boundary_total_pop_col = NULL,
  crs_code = 3005,
  buffer_dist_to_correct_by_meters = 5,
  rand_dist_min = 50,
  rand_dist_max = 200,
  rand_angle_min = 1,
  rand_angle_max = 360,
  min_pop_offset_dis = 1,
  max_pop_offset_dis = 3
)
}
\arguments{
\item{sf_point_data}{a point object that requires offsetting}

\item{sf_boundary}{a polygon object required to constrain offsetting within original boundary}

\item{sf_boundary_id_col}{id column from polygon object used to track join to point object}

\item{sf_boundary_id_col_new_name}{customize output column name of id column from polygon object}

\item{sf_boundary_total_pop_col}{highly encouraged, input column name from polygon object containing total population}

\item{crs_code}{any meter based crs compatible for BC}

\item{buffer_dist_to_correct_by_meters}{distance to buffer polygon boundaries towards inside ensuring offsetting remains within original polygon}

\item{rand_dist_min}{if population data not provided, min value for random distance range to be sampled}

\item{rand_dist_max}{if population data not provided, max value for random distance range to be sampled}

\item{rand_angle_min}{if population data not provided, min value for random angle range to be sampled}

\item{rand_angle_max}{if population data not provided, max value for random angle range to be sampled}

\item{min_pop_offset_dis}{set a minimum offset for population based offsetting (1 to 2 times the average distance is standard)}

\item{max_pop_offset_dis}{set a maximum offset for population based offsetting (3 to 5 times the average distance is standard)}
}
\value{
Returns sf point object where geometry of points have been offset from original input.

Output contains several new columns showing how offsetting occurred, including: the original x/y coordinates, offset x/y coordinates, original boundary id, offset boundary id and the corrected boundary id.

Corrected x/y coordinates ensures any points that were offset beyond their original polygon boundary line (ex placed in water or a neighbouring boundary) are nudged back into their original boundary.

Also prints a summary in console.
}
\description{
Offsets point data to maintain confidentiality.
Designed for BC data; default crs output is 3005 (NAD 1983 / BC Albers); can input other meter based crs like 26910 (NAD83 / UTM zone 10).
Acknowledgement that BCCDC provided resources and offsetting (geomasking) methodology.

Offsetting methodology:

    x offset formula: x' = x0 + RandDist * cos(RandAngle)

    y offset formula: y' = y0 + RandDist * sin(RandAngle)


Key Background:

code relies on sf package for spatial data manipulation.

A point and polygon object in a meters based crs compatible for BC data (crs 3005 or 26910) is required. The polygon boundary layer ensures offset points remain within the area of interest (example, use bcmaps::health_chsa())

Important: if possible, in the polygon boundary layer, include a column with total population for each boundary (ex link to BC Stats' PEOPLE data).

Offsetting is most appropriately done when using the average distance between people within a given area. This is estimated by the square root of the inverse of population density.
From this, we generate a minimum offset (1 to 2 times the average distance, 1 is used in this function) and a maximum offset (3 to 5 times the average distance, 3 is used in this function)

Note: During data preparation, before using this function, consider excluding cases in low population areas.
}
\examples{
\dontrun{
  offset_points_within_boundary(
    sf_point_data = bcmaps::bc_cities(),
    sf_boundary = bcmaps::health_chsa(),
    sf_boundary_id_col = "cmnty_hlth_serv_area_code",
    sf_boundary_id_col_new_name = "chsa",
    sf_boundary_total_pop_col = "chsa_population_census"
    )
  }

}
