% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_point_within_boundary.R
\name{sf_point_within_boundary}
\alias{sf_point_within_boundary}
\title{sf_point_within_boundary}
\usage{
sf_point_within_boundary(
  sf_point_data,
  sf_boundary_data,
  sf_boundary_id_col,
  sf_boundary_id_col_new_name = "boundaryid"
)
}
\arguments{
\item{sf_point_data}{point data of interest}

\item{sf_boundary_data}{boundary of interest joining to}

\item{sf_boundary_id_col}{boundary id col desired to identify boundary assigned to point based on current geometry}

\item{sf_boundary_id_col_new_name}{customize output column name of id column from polygon object}
}
\value{
returns an sf object with a new column showing which boundary id the point falls inside
}
\description{
Point Within Boundary

Used in offset_points_within_boundary(); identifies which boundaries the points fall within.
The function is used to flag offset points that land outside their original boundary polygon and correct their placement back into the original polygon
}
\examples{
\dontrun{
  offset_points_within_boundary(postal_code, bcmaps::health_chsa, "cmnty_hlth_serv_area_code", "chsa", "total_pop_chsa")
  }

}
\keyword{internal}
