% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_offset_corrections.R
\name{summary_offset_corrections}
\alias{summary_offset_corrections}
\title{summary_offset_corrections}
\usage{
summary_offset_corrections(corrected_offset_data, sf_boundary_id_col_new_name)
}
\arguments{
\item{corrected_offset_data}{an offset point sf object that requires adjustments due to offset point falling outside their original boundary polygon}

\item{sf_boundary_id_col_new_name}{beginning part id col being used to track join (ex: "boundaryid")}
}
\value{
returns an sf point object with a new column flagging rows as TRUE if still in original polygon boundary or FALSE if a mismatch still occurs. Also prints summary in console
}
\description{
Summarize Success of Corrected Offset Points

Used in offset_points_within_boundary(); summarizes success of corrected offsetting
}
\examples{
\dontrun{
  offset_points_within_boundary(postal_code, bcmaps::health_chsa, "cmnty_hlth_serv_area_code", "chsa", "total_pop_chsa")
  }

}
\keyword{internal}
