% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_to_crs.R
\name{transform_to_crs}
\alias{transform_to_crs}
\title{transform_to_crs}
\usage{
transform_to_crs(sf_object, desired_crs_code = 3005)
}
\arguments{
\item{sf_object}{sf object's crs to evaluate and transform if needed}

\item{desired_crs_code}{default 3005; any BC meters based crs will work}
}
\value{
returns an sf object
}
\description{
Transform CRS Code

Used in offset_points_within_boundary(); ensures both point and polygon sf objects are in the same crs
}
\examples{
\dontrun{
  offset_points_within_boundary(
    sf_point_data = bcmaps::bc_cities(),
    sf_boundary = bcmaps::health_chsa(),
    sf_boundary_id_col = "cmnty_hlth_serv_area_code",
    sf_boundary_id_col_new_name = "chsa",
    sf_boundary_total_pop_col = "chsa_population_census"
    )
  }

}
\keyword{internal}
