% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ave_dist_btw_ppl.R
\name{ave_dist_btw_ppl}
\alias{ave_dist_btw_ppl}
\title{ave_dist_btw_ppl}
\usage{
ave_dist_btw_ppl(
  sf_boundary,
  total_pop_col,
  min_offset_dis = 1,
  max_offset_dis = 3
)
}
\arguments{
\item{sf_boundary}{sf boundary with a column containing population data}

\item{total_pop_col}{name of column containing population data}

\item{min_offset_dis}{set a minimum offset (1 to 2 times the average distance is standard)}

\item{max_offset_dis}{set a maximum offset (3 to 5 times the average distance is standard)}
}
\value{
returns an sf boundary object with area, population density, and min/max average distance to offset by
}
\description{
Average Distance between People in Each Boundary Polygon

Used in offset_points_within_boundary(); calculates polygon area, population density, and min/max average distance values to offset by
}
\examples{
\dontrun{
  offset_points_within_boundary(postal_code, bcmaps::health_chsa, "cmnty_hlth_serv_area_code", "chsa", "total_pop_chsa")
  }

}
\keyword{internal}
