% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crs_cs_units.R
\name{get_crs_cs_units}
\alias{get_crs_cs_units}
\title{get_crs_cs_units}
\usage{
get_crs_cs_units(sf_obj)
}
\arguments{
\item{sf_obj}{sf object to identify crs' units used}
}
\value{
crs unit inputted; when used in offset_points_within_boundary(), if in meters, runs the rest of the function, otherwise returns an error in console
}
\description{
Check crs' cs unit of measure for x/y coordinates

Used in function offset_points_within_boundary() to flag if a meters based x/y unit is used or not
}
\examples{
\dontrun{
  offset_points_within_boundary(
    sf_point_data = bcmaps::bc_cities(),
    sf_boundary = bcmaps::health_chsa(),
    sf_boundary_id_col = "cmnty_hlth_serv_area_code",
    sf_boundary_id_col_new_name = "chsa",
    sf_boundary_total_pop_col = "chsa_population_census"
    )
  }

}
\keyword{internal}
