% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_is_point_object.R
\name{is_point_sf}
\alias{is_point_sf}
\title{Stop Flags - Is Point Object}
\usage{
is_point_sf(sf_point_object)
}
\arguments{
\item{sf_point_object}{point object}
}
\value{
stops function if not a point/multipoint object
}
\description{
built in checks that stop function if not met
}
\examples{
\dontrun{
  offset_points_within_boundary(
    sf_point_data = bcmaps::bc_cities(),
    sf_boundary = bcmaps::health_chsa(),
    sf_boundary_id_col = "cmnty_hlth_serv_area_code",
    sf_boundary_id_col_new_name = "chsa",
    sf_boundary_total_pop_col = "chsa_population_census"
    )
  }

}
\keyword{internal}
