% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_is_polygon_object.R
\name{is_polygon_sf}
\alias{is_polygon_sf}
\title{Stop Flags - Is Poylgon Object}
\usage{
is_polygon_sf(sf_polygon_object)
}
\arguments{
\item{sf_polygon_object}{polygon object}
}
\value{
stops function if not a polygon/multipolygon object
}
\description{
built in checks that stop function if not met
}
\examples{
\dontrun{
  offset_points_within_boundary(
    sf_point_data = bcmaps::bc_cities(),
    sf_boundary = bcmaps::health_chsa(),
    sf_boundary_id_col = "cmnty_hlth_serv_area_code",
    sf_boundary_id_col_new_name = "chsa",
    sf_boundary_total_pop_col = "chsa_population_census"
    )
  }

}
\keyword{internal}
